fix windows bug.

+++ utils/loss.py
@@ -7,7 +7,7 @@ class MaskedCrossEntropyLoss(nn.Module):
         self.cross_entropy = nn.CrossEntropyLoss(reduction='none')
 
     def forward(self, res, gt, mask):
-        loss = self.cross_entropy(res, gt)
+        loss = self.cross_entropy(res, gt.long())
         loss = loss * mask
         sample_loss = loss.sum(dim = [1, 2]) / mask.sum(dim = [1, 2])
         mean_batch_loss = torch.mean(sample_loss)
